+++ CHANGELOG.md
@@ -1,5 +1,7 @@
-### 0.3.0 | 2019-09-06
+### 0.3.0 | 2019-09-11
 
+- Add `-f` flag to `npx vscode-dts master -f` for forced removal of conflicting types.
+- When `@types/vscode` is present, remove `@types/vscode` folder entirely instead of just removing `@types/vscode/index.d.ts`.
 - Allow `npx vscode-dts dev <git-tag | git-branch>`. [#2](https://github.com/microsoft/vscode-dts/issues/2).
 
 ### 0.2.0 | 2019-06-25